refactor: removed column chart
