+++ projects/safe/src/lib/components/ui/bar-chart/bar-chart.component.html
@@ -1,54 +1,9 @@
-<!-- <kendo-chart #chart [seriesColors]="options && options.palette">
-  <kendo-chart-value-axis>
-    <kendo-chart-value-axis-item [min]="min" [max]="max">
-    </kendo-chart-value-axis-item>
-  </kendo-chart-value-axis>
-  <kendo-chart-title
-    *ngIf="title"
-    [visible]="title.visible"
-    [text]="title.text"
-    [position]="title.position"
-    [font]="title.font"
-    [color]="title.color"
-  >
-  </kendo-chart-title>
-  <kendo-chart-legend
-    *ngIf="legend"
-    [visible]="legend.visible"
-    [position]="legend.position"
-    [orientation]="legend.orientation"
-  >
-  </kendo-chart-legend>
-  <kendo-chart-tooltip format="{0}"></kendo-chart-tooltip>
-  <kendo-chart-series>
-    <kendo-chart-series-item
-      *ngFor="let serie of series"
-      type="bar"
-      field="field"
-      categoryField="category"
-      [stack]="stack"
-      [gap]="gap"
-      [spacing]="spacing"
-      [data]="serie.data"
-      [name]="serie.name || ''"
-      [color]="serie.color"
-    >
-      <kendo-chart-series-item-labels
-        *ngIf="labelContent"
-        position="outsideEnd"
-        color="#000"
-        [content]="labelContent"
-      ></kendo-chart-series-item-labels>
-    </kendo-chart-series-item>
-  </kendo-chart-series>
-</kendo-chart> -->
 <div class="h-full p-2">
   <canvas
     baseChart
     [data]="chartData"
     [options]="chartOptions"
     [type]="chartType"
-    (chartHover)="chartHovered($event)"
-    (chartClick)="chartClicked($event)"
+    [plugins]="plugins"
   ></canvas>
 </div>
