--- projects/safe/src/lib/components/ui/column-chart/column-chart.component.html
@@ -1,9 +1,0 @@
-<div class="h-full p-2">
-  <canvas
-    baseChart
-    [data]="chartData"
-    [options]="chartOptions"
-    [type]="chartType"
-    [plugins]="plugins"
-  ></canvas>
-</div>
