--- projects/safe/src/lib/components/ui/column-chart/column-chart.component.scss
@@ -1,3 +1,0 @@
-.k-chart {
-  height: 100%;
-}
