--- projects/safe/src/lib/components/ui/column-chart/public-api.ts
@@ -1,2 +1,0 @@
-export * from './column-chart.component';
-export * from './column-chart.module';
