+++ projects/safe/src/lib/components/widgets/chart/chart.component.html
@@ -40,17 +40,19 @@
       [series]="series"
       [options]="options"
     ></safe-donut-chart>
-    <safe-column-chart
+    <safe-bar-chart
       #chartWrapper
+      orientation="vertical"
       class="widget-content"
       *ngIf="settings.chart.type === 'column'"
       [title]="settings.chart.title"
       [legend]="settings.chart.legend"
       [series]="series"
       [options]="options"
-    ></safe-column-chart>
+    ></safe-bar-chart>
     <safe-bar-chart
       #chartWrapper
+      orientation="horizontal"
       class="widget-content"
       *ngIf="settings.chart.type === 'bar'"
       [title]="settings.chart.title"
