+++ projects/safe/src/lib/components/widgets/chart/chart.module.ts
@@ -4,7 +4,6 @@ import { SafeChartComponent } from './chart.component';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { SafeButtonModule } from '../../ui/button/button.module';
 import { SafeDonutChartModule } from '../../ui/donut-chart/donut-chart.module';
-import { SafeColumnChartModule } from '../../ui/column-chart/column-chart.module';
 import { SafePieChartModule } from '../../ui/pie-chart/pie-chart.module';
 import { SafeLineChartModule } from '../../ui/line-chart/line-chart.module';
 import { SafeBarChartModule } from '../../ui/bar-chart/bar-chart.module';
@@ -25,7 +24,6 @@ import { TranslateModule } from '@ngx-translate/core';
     SafeDonutChartModule,
     SafePieChartModule,
     SafeLineChartModule,
-    SafeColumnChartModule,
     SafeBarChartModule,
     SafeIconModule,
     TranslateModule,
